+++ R/calc_earthtide.R
@@ -9,10 +9,10 @@
 #' @param method One or more of "gravity", 
 #'     "tidal_potential", "tidal_tilt", "vertical_displacement", 
 #'     "horizontal_displacement", "n_s_displacement", "e_w_displacement",
-#'     "vertical_strain", "areal_strain", "volume_strain", or "ocean_tides",
-#'     "pole_tide", "lod_tide". The pole tide and lod_tide are used in predict
-#'     mode even if do_predict is FALSE. More than one value can only be used
-#'     if do_predict == TRUE.
+#'     "vertical_strain", "areal_strain", "volume_strain", "horizontal_strain",
+#'     or "ocean_tides", "pole_tide", "lod_tide". The pole tide and lod_tide 
+#'     are used in predict mode even if do_predict is FALSE. More than one value
+#'     can only be used if do_predict == TRUE.
 #' @param astro_update Integer that
 #'     determines how often to phases are updated in number of samples. Defaults
 #'     to 1 (every sample), but speed gains are realized with larger values.
@@ -40,6 +40,7 @@
 #'     following columns: datetime, ddt, ut1_utc, lod, x, y, dx, dy
 #' @param return_matrix Return a matrix of tidal values instead of data.frame. 
 #'     The datetime column will not be present in this case (logical).
+#' @param scale Scale results when do_pedict is FALSE
 #' @param ... Currently not used.
 #'
 #' @return data.frame of tidal results
@@ -76,6 +77,7 @@ calc_earthtide <- function(utc,
                       catalog = 'ksm04',
                       eop = NULL,
                       return_matrix = FALSE,
+                      scale = TRUE,
                       ...){
   
   
@@ -124,13 +126,15 @@ calc_earthtide <- function(utc,
       if(return_matrix) {
         return(et$analyze(method = method[i], 
                           astro_update = astro_update, 
-                          return_matrix = return_matrix)
+                          return_matrix = return_matrix, 
+                          scale = scale)
         )
       } else {
         
         et$analyze(method = method[i], 
                    astro_update = astro_update, 
-                   return_matrix = return_matrix)
+                   return_matrix = return_matrix,
+                   scale = scale)
       }
       
     }