+++ tests/testthat/test-earthtide.R
@@ -1,7 +1,7 @@
 context("test-earthtide")
 
 test_that("earthtide works", {
-  
+
   tms <- as.POSIXct('1990-01-01', tz = 'UTC') + c(0, 3600)
   
   wave_groups = data.frame(start = 0, end = 8)
@@ -18,8 +18,10 @@ test_that("earthtide works", {
   et$predict(method = 'tidal_potential', astro_update = 1L)
   et$predict(method = 'tidal_tilt', astro_update = 1L)
   et$predict(method = 'vertical_displacement', astro_update = 1L)
+  et$predict(method = 'horizontal_displacement', astro_update = 1L)
   et$predict(method = 'vertical_strain', astro_update = 1L)
   et$predict(method = 'areal_strain', astro_update = 1L)
+  et$predict(method = 'horizontal_strain', astro_update = 1L)
   et$predict(method = 'volume_strain', astro_update = 1L)
   et$predict(method = 'ocean_tides', astro_update = 1L)
   
@@ -27,9 +29,11 @@ test_that("earthtide works", {
   expect_equal(tide$tidal_potential, c(1.422, 1.890), tolerance = 0.001)
   expect_equal(tide$tidal_tilt, c(-17.845, -20.578), tolerance = 0.001)
   expect_equal(tide$vertical_displacement, c(87.822, 110.715), tolerance = 0.001)
+  expect_equal(tide$horizontal_displacement, c(-46.14956, 53.40266), tolerance = 2) # from solidearthtide
   expect_equal(tide$vertical_strain, c(-5.533, -6.964), tolerance = 0.001)
   expect_equal(tide$areal_strain,  c(16.600, 20.893), tolerance = 0.001)
   expect_equal(tide$volume_strain, c(11.067, 13.929), tolerance = 0.001)
+  expect_equal(tide$horizontal_strain, c(7.958, 11.823), tolerance = 0.001)
   expect_equal(tide$ocean_tides, c(146.327, 183.187), tolerance = 0.01)
   
   # expect_equal(et$predict(method = 'horizontal_displacement', astro_update = 1L)[1:2],